helpo-api/requirements.txt
awscli==1.18.41
numpy==1.18.2
django-debug-toolbar==1.9.1
django-debug-toolbar-request-history==0.0.7
fixmydjango==0.3.2
django==1.11.28
celery==4.2.1
redis==2.10.6
django-js-reverse===0.9.1
djangorestframework==3.9.1
django-import-export==1.0.1
django-cors-headers==2.4.0
django-rest-knox==3.1.5
django-model-utils==3.1.2
django-rest-swagger==2.2.0
python-decouple==3.1
psycopg2==2.7.5
django-log-request-id==1.3.2
dj-database-url==0.5.0
gunicorn==19.9.0
whitenoise==3.3.1
ipython==6.4.0
requests==2.20.0
pandas==0.23.4
sklearn==0.0
boto3==1.9.28
coverage==4.5.1
surprise==0.1
dacite==1.4.0
Inject==3.3.2
pytest==3.6.0
pytest-django==3.2.1